# [kvimd] Add test for writing only once