- [x] Remove comment after benchmarking it
- [x] [kvimd] Add test for writing key only once (rewriting it makes it not override)