@Viq111 2018-12-18
<details>
 <summary>Benchmarks results </summary>

```gf
name               old time/op     new time/op    delta
HashDiskWrite-2       5.32µs ± 5%    5.26µs ±14%    ~     (p=0.905 n=4+5)
HashDiskRead-2        1.09µs ± 1%    1.12µs ± 3%  +2.65%  (p=0.008 n=5+5)
KvimdRandbo-2          344ns ± 0%     347ns ± 0%  +0.75%  (p=0.016 n=5+4)
KvimdWrite-2          5.86µs ± 2%    5.85µs ± 1%    ~     (p=1.000 n=5+5)
KvimdReadSame-2        151ns ±20%     152ns ±20%    ~     (p=0.849 n=5+5)
KvimdReadRandom-2    8.36µs ±255%  11.40µs ±172%    ~     (p=1.000 n=5+5)
ValuesDiskSet-2       34.3ns ± 1%    34.7ns ± 2%    ~     (p=0.151 n=5+5)
ValuesDiskGet-2       33.0ns ± 2%    32.9ns ± 1%    ~     (p=0.952 n=5+5)

name               old speed       new speed      delta
HashDiskWrite-2     4.51MB/s ± 5%  4.59MB/s ±15%    ~     (p=0.794 n=4+5)
HashDiskRead-2      21.9MB/s ± 1%  21.4MB/s ± 3%  -2.54%  (p=0.008 n=5+5)
KvimdRandbo-2        336MB/s ± 0%   333MB/s ± 0%  -0.85%  (p=0.008 n=5+5)
KvimdWrite-2        19.8MB/s ± 2%  19.8MB/s ± 1%    ~     (p=1.000 n=5+5)
KvimdReadSame-2      709MB/s ±64%   709MB/s ±64%    ~     (p=0.841 n=5+5)
KvimdReadRandom-2  47.0MB/s ±100%  48.5MB/s ±97%    ~     (p=1.000 n=5+5)
ValuesDiskSet-2      233MB/s ± 1%   231MB/s ± 2%    ~     (p=0.222 n=5+5)
ValuesDiskGet-2      242MB/s ± 2%   243MB/s ± 1%    ~     (p=0.841 n=5+5)
```